refactor: move previewPanel code and add tests
